[api] Implements accepted_at time for Review objects
The Review#accepted_at timestamp is calculated from the HistoryElement::ReviewAccepted#created_at timestamp.
This object gets created when a Review gets accepted.
However, if a Group review (e.g. maintenance group) gets assigned to a user, the accepted_at time is taken from the assigned user review
and not from the group review.
We calculate that from the generated HistoryElements. However, we need to rely on the comments in the HistoryElements for now.
It should be considered to introduce associations between the HistoryElements / Reviews to make this more bulletproof.